Quick cleanup, and added pedinfo command for help
