+++ test-gametype/client/Utils.ts
@@ -1,14 +1,3 @@
-export type PlayerInfo = {
-  id: number;
-  serverId: number;
-  name: string;
-  ped: number;
-};
-
-export type PlayerIdMap = Record<number, PlayerInfo>;
-export type PlayerPedMap = Record<number, PlayerInfo>;
-export type PlayerNameMap = Record<string, PlayerInfo>;
-
 export type Vector3 = {
   x: number;
   y: number;
